V792 (QDC): deal with chopped events
Under high load, V792 readout may result in splitted events, with one
half of the event being in one readout buffer, and the other half being
in the subsequent readout buffer. This problem has been observed and
solved for V1290 (TDC). Generalize this solution by moving its code to
Digitizer and apply it to V792.